requirements.txt
pipecat-ai
aiohttp
python-dotenv
loguru
piper-tts
onnxruntime
flask
pipecat-ai[local-smart-turn-v3]
pipecat-ai[aws]
boto3
pipecat-ai[daily]
pipecat-ai[runner]
pipecat-ai[webrtc]
pipecat-ai[tracing]
fastapi
websockets
faster-whisper